Image histogram, integrate NoIR cam

+++ install.sh
@@ -28,22 +28,26 @@ armv6l | armv7l | armhf)
   ;;
 esac
 
-echo "Detected architecture: $ARCH_RAW -> using '$ARCH'"
+if [ -x "src/platesolving/astap_cli" ]; then
+  echo "Already installed astap_cli, skipping."
+else
+  echo "Detected architecture: $ARCH_RAW -> using '$ARCH'"
 
-URL="https://sourceforge.net/projects/astap-program/files/linux_installer/astap_command-line_version_Linux_${ARCH}.zip/download"
-OUTPUT="astap_cli.zip"
+  URL="https://sourceforge.net/projects/astap-program/files/linux_installer/astap_command-line_version_Linux_${ARCH}.zip/download"
+  OUTPUT="astap_cli.zip"
 
-echo "Downloading from: $URL"
-curl -L -o "$OUTPUT" "$URL"
+  echo "Downloading from: $URL"
+  curl -L -o "$OUTPUT" "$URL"
 
-echo "Saved as: $OUTPUT"
-echo "Unzipping"
-sudo apt install -y unzip
-unzip astap_cli.zip
-rm -rf astap_cli.zip
-chmod +x astap_cli
-mv astap_cli src/platesolving/astap_cli
-echo "ASTAP Installed at src/platesolving/astap_cli"
+  echo "Saved as: $OUTPUT"
+  echo "Unzipping"
+  sudo apt install -y unzip
+  unzip astap_cli.zip
+  rm -rf astap_cli.zip
+  chmod +x astap_cli
+  mv astap_cli src/platesolving/astap_cli
+  echo "ASTAP Installed at src/platesolving/astap_cli"
+fi
 
 echo "Installing dependencies"
 sudo apt install -y build-essential libcap-dev python3-dev python3-libcamera python3-kms++
@@ -52,4 +56,3 @@ echo "Setting up python venv"
 python3 -m venv --system-site-packages .venv
 source .venv/bin/activate
 pip install -r requirements.txt --prefer-binary
-
